bug_fix: Fix uint32_t overflow for timer arithmetic
Intentionally left out some portions since overflow isn't realistically going to occur unless someone keeps the computer on for a very long while or at a very high frequency.

+++ src/drivers/pit.c
@@ -60,7 +60,7 @@ uint32_t ticks_to_ms(uint32_t ticks)
     if (timer_freq == 0)
         return 0;
 
-    return (ticks * 1000) / timer_freq;
+    return (uint32_t)((uint64_t)ticks * 1000) / timer_freq;
 }
 
 uint32_t get_timer_freq(void)
